placed all detection functions into seperate module
cleaned up main module and created detection module for single process detection functions. poderpools has the default multiprocess functions
